refactor: schoolIds, departments 출처 묶기
